+++ super-mario-3-se102/ProcessingUtils.cpp
@@ -59,7 +59,7 @@ std::string ProcessingUtils::Trim(const std::string& str)
 {
 	int start = -1;
 	int end = -1;
-	for (size_t i = 0; i < str.size(); i++)
+	for (int i = 0; i < static_cast<int>(str.size()); i++)
 		if (str[i] != ' ') {
 			start = i;
 			break;
